feat: revert organization changes for server actions
