+++ src/actions/reactions.ts
@@ -1,7 +1,11 @@
 import { unstable_noStore as noStore } from 'next/cache';
 import { cache } from 'react';
 
-import { db, type ReactionName } from '@/lib/planetscale';
+import {
+  db,
+  type ReactionName,
+  type ReactionsCounters,
+} from '@/lib/planetscale';
 
 import { canRunAction } from './utils';
 
@@ -28,20 +32,25 @@ export const incrementReaction = cache(
   },
 );
 
-export const recordView = cache(async (slug: string) => {
-  if (!canRunAction) return;
+export const getReactions = async (
+  slug: string,
+): Promise<ReactionsCounters> => {
   noStore();
   try {
     const data = await db
       .selectFrom('counters')
       .where('slug', '=', slug)
-      .select(['views'])
+      .select(['likes', 'loves', 'awards', 'bookmarks'])
       .execute();
-
-    const views = !data.length ? 0 : Number(data[0].views);
-    db.insertInto('counters')
-      .values({ slug, views: 1 })
-      .onDuplicateKeyUpdate({ views: views + 1 })
-      .execute();
-  } catch (e) {}
-});
+    const [counters] = data;
+    if (!counters) return {};
+    return {
+      likes: counters.likes || 0,
+      loves: counters.loves || 0,
+      awards: counters.awards || 0,
+      bookmarks: counters.bookmarks || 0,
+    };
+  } catch (e) {
+    return {};
+  }
+};
