+++ src/actions/stars.tsx
