+++ src/actions/views.ts
@@ -1,29 +1,27 @@
 import { unstable_noStore as noStore } from 'next/cache';
+import { cache } from 'react';
 
-import { db, type ReactionsCounters } from '@/lib/planetscale';
+import { db } from '@/lib/planetscale';
 
-export const getReactions = async (
-  slug: string,
-): Promise<ReactionsCounters> => {
+import { canRunAction } from './utils';
+
+export const recordView = cache(async (slug: string) => {
+  if (!canRunAction) return;
   noStore();
   try {
     const data = await db
       .selectFrom('counters')
       .where('slug', '=', slug)
-      .select(['likes', 'loves', 'awards', 'bookmarks'])
+      .select(['views'])
       .execute();
-    const [counters] = data;
-    if (!counters) return {};
-    return {
-      likes: counters.likes || 0,
-      loves: counters.loves || 0,
-      awards: counters.awards || 0,
-      bookmarks: counters.bookmarks || 0,
-    };
-  } catch (e) {
-    return {};
-  }
-};
+
+    const views = !data.length ? 0 : Number(data[0].views);
+    db.insertInto('counters')
+      .values({ slug, views: 1 })
+      .onDuplicateKeyUpdate({ views: views + 1 })
+      .execute();
+  } catch (e) {}
+});
 
 export const getViews = async (slug: string): Promise<number> => {
   noStore();
